refactor: Add simulation in pharmacy + test API breaking
